requirements.txt
fastapi
python-fasthtml
bcrypt